// AngularMain/src/app/modules/models/ServiceSetup/ServiceSetupDto.ts
export interface ServiceSetupDto {
    tenentId: number;
    serviceId: number;
    masterServiceId: string[];
    serIdbyUser: number;
    serviceName1: string;
    serviceName2: string;
    serviceType: number;
    serviceTypeName:string;
    serviceSubType: number;
    allowSponser: number;
    allowedNonEmployes?: boolean;
    minMonthsService: number;
    minInstallment: number;
    maxInstallment: number;
    frozen?: boolean;
    previousEmployees?: boolean;
    serApproval1: string;
    approvalBy1: string;
    approvedDate1: Date;
    serApproval2: string;
    approvalBy2: string;
    approvedDate2: Date;
    serApproval3: string;
    approvalBy3: string;
    approvedDate3: Date;
    serApproval4: string;
    approvalBy4: string;
    approvedDate4: Date;
    serApproval5: string;
    approvalBy5: string;
    approvedDate5: Date;
    finalApproval: string;
    remarks: string;
    keyword: string;
    loanAct: string;
    hajjAct: string;
    persLoanAct: string;
    consumerLoanAct: string;
    otherAct1: string;
    otherAct2: string;
    otherAct3: string;
    otherAct4: string;
    otherAct5: string;
    sortBy: number;
    active: string;
    crupId: number;
    userid: string;
    entrydate: Date;
    entrytime: Date;
    updttime: Date;
    uploadDate: Date;
    uploadby: string;
    syncDate: Date;
    syncby: string;
    synId: number;
    allowDiscountPer?: boolean;
    discountType:number;
    allowDiscountAmount?: number;
    allowDiscountDefault?: boolean;
    englishHTML?: string;
    arabicHTML?: string;
    electronicForm1?: string;
    electronicForm1URL?: string;
    electronicForm2?: string;
    electronicForm2URL?: string;
    contractType?:string
}